DISPATCH-1311: move qdr_delivery_t API to its own source files
This patch merely moves code and does some renaming of functions.  No
changes to actual functionality have occurred - it is simply a
reorganization of code layout.

This closes #489
