+++ include/qpid/dispatch/router_core.h
@@ -660,8 +660,6 @@ qdr_delivery_t *qdr_link_deliver_to(qdr_link_t *link, qd_message_t *msg,
 qdr_delivery_t *qdr_link_deliver_to_routed_link(qdr_link_t *link, qd_message_t *msg, bool settled,
                                                 const uint8_t *tag, int tag_length,
                                                 uint64_t disposition, pn_data_t* disposition_state);
-qdr_delivery_t *qdr_deliver_continue(qdr_core_t *core, qdr_delivery_t *delivery);
-
 int qdr_link_process_deliveries(qdr_core_t *core, qdr_link_t *link, int credit);
 
 void qdr_link_flow(qdr_core_t *core, qdr_link_t *link, int credit, bool drain_mode);
@@ -694,34 +692,6 @@ void qdr_connection_handlers(qdr_core_t             *core,
                              qdr_link_deliver_t         deliver,
                              qdr_delivery_update_t      delivery_update,
                              qdr_connection_close_t     conn_close);
-
-/**
- ******************************************************************************
- * Delivery functions
- ******************************************************************************
- */
-void qdr_delivery_update_disposition(qdr_core_t *core, qdr_delivery_t *delivery, uint64_t disp,
-                                     bool settled, qdr_error_t *error, pn_data_t *ext_state, bool ref_given);
-
-void qdr_delivery_set_context(qdr_delivery_t *delivery, void *context);
-void *qdr_delivery_get_context(qdr_delivery_t *delivery);
-qdr_link_t *qdr_delivery_link(const qdr_delivery_t *delivery);
-void qdr_delivery_incref(qdr_delivery_t *delivery, const char *label);
-void qdr_delivery_decref(qdr_core_t *core, qdr_delivery_t *delivery, const char *label);
-void qdr_delivery_tag(const qdr_delivery_t *delivery, const char **tag, int *length);
-qd_message_t *qdr_delivery_message(const qdr_delivery_t *delivery);
-qdr_error_t *qdr_delivery_error(const qdr_delivery_t *delivery);
-bool qdr_delivery_presettled(const qdr_delivery_t *delivery);
-void qdr_delivery_write_extension_state(qdr_delivery_t *dlv, pn_delivery_t* pdlv, bool update_disposition);
-bool qdr_delivery_send_complete(const qdr_delivery_t *delivery);
-bool qdr_delivery_tag_sent(const qdr_delivery_t *delivery);
-void qdr_delivery_set_tag_sent(const qdr_delivery_t *delivery, bool tag_sent);
-bool qdr_delivery_receive_complete(const qdr_delivery_t *delivery);
-void qdr_delivery_set_disposition(qdr_delivery_t *delivery, uint64_t disposition);
-uint64_t qdr_delivery_disposition(const qdr_delivery_t *delivery);
-void qdr_delivery_set_aborted(const qdr_delivery_t *delivery, bool aborted);
-bool qdr_delivery_is_aborted(const qdr_delivery_t *delivery);
-void qdr_delivery_add_num_closed_receivers(qdr_delivery_t *delivery);
 
 /**
  ******************************************************************************
